pack200 is no longer supported: it was encountered very seldomly and the unpack200 tool will be deprecated soon (and pack200 also no longer seems to work, making it hard to create test files)
