+++ shell.nix
@@ -39,7 +39,6 @@ pkgs.mkShell {
     lz4
     mailcap
     ncompress
-    openjdk8
     openssl
     my-python
     protobuf
