+++ src/bangsignatures.py
@@ -252,7 +252,6 @@ signaturetofunction = {
     'swf_lzma': bangmedia.unpack_swf,
     'certificate': bangunpack.unpack_certificate,
     'pdf': bangmedia.unpack_pdf,
-    'pack200': bangunpack.unpack_pack200,
     'zim': bangunpack.unpack_zim,
     'sqlite3': bangunpack.unpack_sqlite,
     'trx': bangunpack.unpack_trx,
