+++ src/bangunpack.py
@@ -3962,86 +3962,6 @@ def extract_certificate(filename_full, scanenvironment, offset):
     return {'status': False, 'error': unpackingerror}
 
 
-# Transform a pack200 file to a JAR file using the unpack200 tool.
-# This will not restore the original JAR file, as pack200 performs all kinds
-# of optimizations, such as removing redundant classes, and so on.
-#
-# https://docs.oracle.com/javase/7/docs/technotes/guides/pack200/pack-spec.html
-#
-# The header format is described in section 5.2
-def unpack_pack200(fileresult, scanenvironment, offset, unpackdir):
-    '''Convert a pack200 file back into a JAR'''
-    filesize = fileresult.filesize
-    filename_full = scanenvironment.unpack_path(fileresult.filename)
-    unpackdir_full = scanenvironment.unpack_path(unpackdir)
-    unpackedfilesandlabels = []
-    labels = []
-    unpackingerror = {}
-    unpackedsize = 0
-    unpackdir_full = scanenvironment.unpack_path(unpackdir)
-
-    # first check if the unpack200 program is actually there
-    if shutil.which('unpack200') is None:
-        unpackingerror = {'offset': offset+unpackedsize, 'fatal': False,
-                          'reason': 'unpack200 program not found'}
-        return {'status': False, 'error': unpackingerror}
-
-    # the unpack200 tool only works on whole files. Finding out
-    # where the file ends is TODO, but if there is data in front
-    # of a valid pack200 file it is not a problem.
-    if offset != 0:
-        # create a temporary file and copy the data into the
-        # temporary file if offset != 0
-        checkfile = open(filename_full, 'rb')
-        temporaryfile = tempfile.mkstemp(dir=scanenvironment.temporarydirectory)
-        os.sendfile(temporaryfile[0], checkfile.fileno(), offset, filesize - offset)
-        os.fdopen(temporaryfile[0]).close()
-        checkfile.close()
-
-    # write unpacked data to a JAR file
-    outfile_rel = os.path.join(unpackdir, "unpacked.jar")
-    outfile_full = scanenvironment.unpack_path(outfile_rel)
-
-    # create the unpacking directory
-    os.makedirs(unpackdir_full, exist_ok=True)
-
-    # then extract the file
-    if offset != 0:
-        p = subprocess.Popen(['unpack200', temporaryfile[1], outfile_full],
-                             stdout=subprocess.PIPE, stderr=subprocess.PIPE,
-                             cwd=unpackdir_full)
-    else:
-        p = subprocess.Popen(['unpack200', filename_full, outfile_full],
-                             stdout=subprocess.PIPE, stderr=subprocess.PIPE,
-                             cwd=unpackdir_full)
-    (outputmsg, errormsg) = p.communicate()
-
-    if offset != 0:
-        os.unlink(temporaryfile[1])
-
-    if p.returncode != 0:
-        # try to remove any files that were possibly left behind
-        try:
-            os.unlink(outfile_full)
-        except:
-            pass
-        unpackingerror = {'offset': offset+unpackedsize, 'fatal': False,
-                          'reason': 'Not a valid pack200 file'}
-        return {'status': False, 'error': unpackingerror}
-
-    unpackedsize = filesize - offset
-
-    if offset == 0 and unpackedsize == filesize:
-        labels.append('pack200')
-
-    unpackedfilesandlabels.append((outfile_rel, []))
-
-    return {'status': True, 'length': unpackedsize, 'labels': labels,
-            'filesandlabels': unpackedfilesandlabels}
-
-unpack_pack200.signatures = {'pack200': b'\xca\xfe\xd0\x0d'}
-
-
 # https://wiki.openzim.org/wiki/ZIM_file_format
 # https://wiki.openzim.org/wiki/ZIM_File_Example
 # Test files: https://wiki.kiwix.org/wiki/Content_in_all_languages
